+++ CHANGELOG.md
@@ -31,6 +31,8 @@
 - [261](https://github.com/vegaprotocol/protos/pull/261) - Add `staking_asset_total_supply` to staking accounts snapshot
 - [272](https://github.com/vegaprotocol/protos/pull/272) - Update snapshot protos to be able to handle the iavl exported nodes
 - [276](https://github.com/vegaprotocol/protos/pull/276) - Remove maturity from future product, tick size from market and trading mode config from market
+- [277](https://github.com/vegaprotocol/protos/pull/277) - Update validation for risk params for log normal risk model
+
 ### 🐛 Fixes
 - [202](https://github.com/vegaprotocol/protos/pull/202) - Replaces Withdrawal Status Cancelled with Rejected which is more accurate.
 - [214](https://github.com/vegaprotocol/protos/pull/214) Rewording of transaction nonce and block_height